# CUDA: use mma PTX instructions for FlashAttention