@sorasoras 2025-02-02
> The CUDA 11.7 compilation problems on Windows seem to be because the `movmatrix` instruction which transposes an 8x8 tile of 16 bit values needs a higher PTX ISA version than the `mma` instruction even though both need Turing or newer in terms of hardware. This is very annoying but since this instruction is not performance-critical I think I can write a workaround either with `__shfl` instructions or by going through shared memory.

Can there be a ptx kernel for Pascal via Dp4a？