@JohannesGaessler 2025-02-02
For Pascal I am not aware of any useful PTX instructions that are not already being used in CUDA code.

As of right now quantized KV caches are converted to floating point numbers for large batch FlashAttention. It would in principle be possible to write a FlashAttention kernel that uses int8 arithmetic instead. This would very likely be faster on Pascal, on Turing or newer it's a bit unclear since it's difficult to get good int8 tensor core utilization with the GGML quantization formats.