@JohannesGaessler 2025-02-02
I pushed a version with a workaround for `movmatrix`. The difference in the total kernel runtime is ~1%, for the end-to-end runtime it should be negligible. I don't know how to query the max. supported PTX version from CUDA code, I made the switch based on CUDA 11 vs. 12; I'll ask about a better way to do this on Friday when I talk to NVIDIA engineers.