@JohannesGaessler 2025-02-02
While going through the PTX ISA documentation I noticed that there is a table that maps PTX versions to CUDA versions. Using that table I could determine that `movmatrix` should be available starting with CUDA 11.8 and I changed the check accordingly.