@JohannesGaessler 2025-02-02
>I wish we could keep the wmma implementation, as this performs decently on mfma/rocmwmma for large batch sizes

For CUDA the implementation is definitely bad and in a vacuum I would not want to invest the effort to maintain it long-term. However, I recognize that you're making valuable contributions to the project and I'm willing to in return maintain the wmma code. I would also be willing to merge an optional compilation argument for rocwmma if I get a pledge from you that you will maintain it (since I currently do not have any suitable hardware for testing, may change in March).

But I must stress that relative to what the hardware would be capable of the performance of a HIP port of my wmma code will always be bad. Ideally we would just have a ROCm implementation. I would be willing to learn ROCm in order to review any related code.