@IMbackK 2025-02-02
Dont get me wrong, its bad, it just performs better than the other attn implementations available in ggml for now. Lets discuss details at a later date, as for now as i understand it its in no intimidate risk of being removed, as it still benefits volta also.