@ggerganov 2025-02-03
@JohannesGaessler After this change, the `ggml-ci` node that runs the CUDA builds is failing because it uses V100:

https://github.com/ggml-org/ci/blob/results/llama.cpp/86/4a0b67a6c8f648c43ce8271f9cb2e12dd5df6e/ggml-4-x86-cuda-v100/stdall#L12402

How to modify the [ggml-ci](https://github.com/ggerganov/llama.cpp/blob/6eecde3cc8fda44da7794042e3668de4af3c32c6/ci/run.sh#L44-L48) script to fix this?